refactor: proxy class to a proxies folder.
