+++ src/common/index.ts
@@ -1,7 +1,7 @@
 export * from './substrate';
-export * from './date-time';
+export * from './proxies/date-time';
 export * from './mailer';
-export * from './process-env';
+export * from './proxies/process-env';
 export * from './polkadot-provider';
 export * from './interceptors';
 export * from './health-indicators';
