+++ src/endpoints/bounty/bounty.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { BountyController } from './bounty.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { DataStakingEvents } from './models/data-staking-events.entity';
-import { DateTimeModule } from '../../common/date-time/date-time.module';
+import { DateTimeModule } from '../../common/proxies/date-time/date-time.module';
 import { DataTokenToDatasetMapping } from './models/data-token-to-dataset-mapping.entity';
 import { GCloudStorageModule } from '@aginix/nestjs-gcloud-storage';
