+++ src/endpoints/cloud-storage/cloud-storage.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { GCloudStorageModule } from '@aginix/nestjs-gcloud-storage';
 import { CloudStorageController } from './cloud-storage.controller';
-import { DateTimeModule } from '../../common/date-time/date-time.module';
+import { DateTimeModule } from '../../common/proxies/date-time/date-time.module';
 
 require('dotenv').config(); // eslint-disable-line
