+++ src/endpoints/verification/verification.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
-import { DateTimeModule } from '../../common/date-time/date-time.module';
-import { ProcessEnvModule } from '../../common/process-env/process-env.module';
+import { DateTimeModule } from '../../common/proxies/date-time/date-time.module';
+import { ProcessEnvModule } from '../../common/proxies/process-env/process-env.module';
 import { RewardModule } from '../../common/reward/reward.module';
 import { SubstrateModule } from '../../substrate/substrate.module';
 import { VerificationController } from './verification.controller';
