+++ src/endpoints/verification/verification.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@nestjs/common';
-import { DateTimeProxy } from '../../common/date-time';
+import { DateTimeProxy } from '../../common/proxies/date-time';
 import { RewardDto } from '../../common/reward/dto/reward.dto';
 import { RewardService } from '../../common/reward/reward.service';
 import { SubstrateService } from '../../substrate/substrate.service';
