+++ test/unit/endpoints/verification/verification.service.spec.ts
@@ -1,7 +1,7 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { dateTimeProxyMockFactory, MockType } from '../../mock';
 import { when } from 'jest-when';
-import { DateTimeProxy } from '../../../../src/common/date-time';
+import { DateTimeProxy } from '../../../../src/common/proxies/date-time';
 import { SubstrateService } from '../../../../src/substrate/substrate.service';
 import { RewardService } from '../../../../src/common/reward/reward.service';
 import { VerificationService } from '../../../../src/endpoints/verification/verification.service';
